fix error прерывистость звука

+++ 6_promwad2_step_by_step/H753_new/Core/Src/Call_Control/callwaiting.cpp
@@ -9,7 +9,7 @@
 #include "rs232.h"
 extern SAI_HandleTypeDef audioTxSai;
 extern uint16_t lastDirectSubject;
-extern int volatile asteriskPressed;
+extern uint8_t volatile asteriskPressed;
 bool asteriskRecall = false;
 extern char msg[50];
 void CallWaiting::handleButton()
